+++ src/CirculationLogList/CirculationLogListContainer.js
@@ -27,6 +27,8 @@ const CirculationLogListContainerComponent = ({ mutator }) => {
 
   const [servicePoints, setServicePoints] = useState();
 
+  const focusRef = React.useRef(null);
+
   useEffect(() => {
     mutator.logEventsServicePoints.GET()
       .then(setServicePoints);
@@ -49,6 +51,8 @@ const CirculationLogListContainerComponent = ({ mutator }) => {
       ...prevLogEvens,
       ...logEventsResponse.logRecords,
     ]));
+
+    return focusRef.current?.focus?.();
   }, []);
 
   const {
@@ -66,6 +70,7 @@ const CirculationLogListContainerComponent = ({ mutator }) => {
       isLoading={isLoading}
       logEvents={logEvents}
       servicePoints={servicePoints}
+      focusRef={focusRef}
     />
   );
 };